requirements.txt
discord.py
discord.py[voice]
pytube
asyncio
python-dotenv